pyproject.toml
[project]
name = "history-guardian"
version = "0.3.0"
description = "GraphRAG-based system for detecting fake news and historical inaccuracies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.31.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "neo4j>=5.17.0",
    "llama-index>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-graph-stores-neo4j>=0.1.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
]
ml = [
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]